+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/model/nip85TrustedAssertions/TrustProviderListState.kt
@@ -18,24 +18,16 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.model.trustedAssertions
+package com.vitorpamplona.amethyst.commons.model.nip85TrustedAssertions
 
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.PrivateTagArrayEventCache
-import com.vitorpamplona.quartz.nip85TrustedAssertions.list.TrustProviderListEvent
-import com.vitorpamplona.quartz.nip85TrustedAssertions.list.serviceProviderSet
-import com.vitorpamplona.quartz.nip85TrustedAssertions.list.serviceProviders
+import com.vitorpamplona.quartz.nip85TrustedAssertions.list.tags.ServiceProviderTag
+import kotlinx.coroutines.flow.StateFlow
 
-class TrustProviderListDecryptionCache(
-    val signer: NostrSigner,
-) {
-    val cachedPrivateLists = PrivateTagArrayEventCache<TrustProviderListEvent>(signer)
-
-    fun cachedServiceProviders(event: TrustProviderListEvent) = cachedPrivateLists.mergeTagListPrecached(event).serviceProviders()
-
-    fun cachedServiceProviderSet(event: TrustProviderListEvent) = cachedPrivateLists.mergeTagListPrecached(event).serviceProviderSet()
-
-    suspend fun serviceProviders(event: TrustProviderListEvent) = cachedPrivateLists.mergeTagList(event).serviceProviders()
-
-    suspend fun serviceProviderSet(event: TrustProviderListEvent) = cachedPrivateLists.mergeTagList(event).serviceProviderSet()
+/**
+ * Interface for trust provider list state.
+ * Used by UserCardsCache for accessing user rank and follower count providers.
+ */
+interface TrustProviderListState {
+    val liveUserRankProvider: StateFlow<ServiceProviderTag?>
+    val liveUserFollowerCount: StateFlow<ServiceProviderTag?>
 }
